Allow copying to clipboard in quiet mode
aiac now supports copying to clipboard in quiet mode too. Due to how
the X11 clipboard works in Linux, the process must remain live after
copying to the clipboard, otherwise the clipboard is cleared. Users
can send aiac to the background if they so wish, and the process will
exit once the clipboard contents changes.