Add support for Amazon Bedrock
This commit introduces support for using Amazon Bedrock as a
backend for code generation. In order to achieve this, the client
structure in libaiac is now based on an interface, with two
implementations. The previous OpenAI implementation is moved to
libaiac/openai, and a Bedrock implementation is introduced at
libaiac/bedrock.

The interface, described in the new libaiac/types package, is
fairly simple, defining a `ListModels`, `DefaultModel`, `Complete`
and `Chat` methods that must be implemented.

These changes are backwards compatible from both the perspective
of a command line usage and library usage, but the major version
is bumped to 4 due to the significance of the change. This means
that users of the library will need to change import statements
to use "v4" instead of "v3".