Add support for the Ollama backend
This commit introduces support for the ollama.ai backend. Ollama is
an open source LLAMA backend meant for local usage. It supports
many different models, many of them related to code generation.

Usage of the ollama backend is available via the `--backend` flag
or `AIAC_BACKEND` environment variables set to `"ollama"`. Ollama
doesn't support authentication currently, so the only related flag
is `--ollama-url` for the API server's URL, but if not used, the
default URL is used (http://localhost:11434/api). With this commit,
`aiac` will not yet support a scenario where the API server is
running behind an authenticating proxy.

Resolves: #77