Support sending extra headers and changing auth header
This commit introduces support for adding extra HTTP headers that
will be sent on every request to backends of type "openai" and
"ollama". This is useful for providers implementing the OpenAI
API but requiring extra headers, such as Portkey.

For openai backends, the header used for authorization can also now
be modified via the configuration setting `auth_header`. Azure
OpenAI uses the "api-key" header instead of "Authorization".
Previously, aiac used this header whenever the URL for the backend
was not the default OpenAI API URL. This is clearly wrong as it's
really only true for Azure OpenAI, and wrong for providers like
Portkey. The new setting is now used instead, but if the URL is
an Azure OpenAI URL, aiac will automatically know to use "api-key"
instead of "Authorization".