Allow conversations to have "previous messages"
This commit modifies the `Chat` method (in a backwards compatible
way) to allow users to start the conversation with previously
exchanged messages, i.e. messages that had already been exchanged
between the user and the assistant. This practically allows
"loading" previous conversations in order to continue them.

To support this, the `Chat` method now takes `types.Message`
values in a variadic way. The messages field of the `Conversation`
type is also exposed now, so that users may save it.