openai: do not force API key, fix error processing
OpenAI backends will no longer require an API key to be provided. This
is because users may wish to use a service such as Portkey with its
Virtual Keys feature, which doesn't require an API key. If an API
key is not provided, an Authorization header is not sent.

Also, error handling in the openai implementation is both fixed and
expanded to support Portkey error format.